server/requirements.txt
channels==3.0.5
channels-redis==3.4.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
hiredis==2.0.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytest-asyncio==0.18.3
pytest-django==4.5.2
redis==4.3.4